+++ CHANGELOG.md
@@ -3791,3 +3791,4 @@
 - [2026-08-20 22:50] feat(state): add async state stream provider
 - [2026-08-20 23:13] refactor(auth): extract reusable atomic widgets from auth
 - [2026-08-21 09:07] perf(ui): apply RepaintBoundary on animated ui widgets
+- [2026-08-21 09:18] fix(forms): handle null assertion failure in forms parser